feat: app에서 훅 분리
