+++ src/hooks/auth/useAuthCheck.ts
@@ -1,0 +1,10 @@
+import { useEffect } from 'react';
+import { useAuthStore } from '@/store/authStore';
+
+export function useAuthCheck() {
+  const { checkAuth } = useAuthStore();
+
+  useEffect(() => {
+    checkAuth();
+  }, [checkAuth]);
+}
