+++ src/hooks/auth/useKakaoInit.ts
@@ -1,0 +1,17 @@
+import { useEffect } from 'react';
+
+export function useKakaoInit() {
+  useEffect(() => {
+    if (typeof window !== 'undefined') {
+      const script = document.createElement('script');
+      script.src = 'https://developers.kakao.com/sdk/js/kakao.js';
+      script.async = true;
+      script.onload = () => {
+        if (window.Kakao && !window.Kakao.isInitialized()) {
+          window.Kakao.init(process.env.NEXT_PUBLIC_KAKAO_JAVASCRIPT_KEY as string);
+        }
+      };
+      document.body.appendChild(script);
+    }
+  }, []);
+}
